src/requirements.txt
django-phonenumber-field==2.2.0
phonenumbers
Django==3.2
Pillow==5.2.0
